In some cases the Wi-Fi Display (WFD) IE was not added to some types of frames. In other cases the WFD IE was not the last IE in the frame. This is fixed.
The Wi-Fi Display (WFD) IE is not included in the following frames.
It is now included in these frames when WFD is enabled:
	P2P Invitation Request
	P2P Invitation Response
	P2P Provision Discovery Response

The WFD IE is not the last IE in the frame in the following frames.
It is now the last IE in the frame.
	P2P GO Negotiation Request
	P2P GO Negotiation Response
	P2P Provision Discovery Request

CRs- Fixed: 342755
(cherry picked from commit de5126a8a2b028b4fdaa24653dcf665f1b9f85fb)

Change-Id: I39e456450789e56ccf1cfc553e90a15a385671d3